backend/pyproject.toml
[project]
name = "ai-agent-backend"
version = "0.1.0"
description = "AI Agent System - Backend Service"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "AI Agent Team" }]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.8.0",

    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.0.0",  # Email validation for Pydantic

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Cache & Queue
    "redis>=5.0.0",
    "celery>=5.3.0",

    # LLM
    "litellm>=1.20.0",
    "tiktoken>=0.5.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",

    # Agent Framework
    "langgraph>=1.0.0",
    "langchain-openai>=1.0.0",  # 用于代码生成器生成的 LLM 节点

    # Vector Database
    "qdrant-client>=1.7.0",
    "chromadb>=1.0.0",

    # Utilities
    "httpx>=0.26.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "python-dotenv>=1.0.0",

    # Code Analysis (for Code-First)
    "libcst>=1.1.0",

    # Evaluation
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    # Type Checking
    "pyright>=1.1.350",
    "mypy>=1.8.0",

    # Linting & Formatting
    "ruff>=0.2.0",
    "pylint>=3.0.0",

    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "faker>=22.0.0",

    # Development
    "pre-commit>=3.6.0",
    "ipython>=8.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "api", "core", "models", "schemas", "services", "db", "utils", "workers"]

# ============================================================================
# Type Checking - Pyright
# ============================================================================
[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportDuplicateImport = true
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/migrations",
    ".venv",
]

# ============================================================================
# Type Checking - MyPy
# ============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langgraph.*",
    "litellm.*",
    "celery.*",
    "qdrant_client.*",
    "chromadb.*",
]
ignore_missing_imports = true

# ============================================================================
# Linting & Formatting - Ruff
# ============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors (includes E402: module level import not at top)
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (import sorting and organization)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",      # line too long - handled by formatter
    "RUF001",    # ambiguous unicode in string (Chinese characters)
    "RUF002",    # ambiguous unicode in docstring (Chinese characters)
    "RUF003",    # ambiguous unicode in comment (Chinese characters)
    "B008",      # function-call-in-default-argument (FastAPI Depends pattern)
    "ARG001",    # unused-function-argument (common in interfaces)
    "ARG002",    # unused-method-argument (common in interfaces)
    "BLE001",    # broad-exception-caught (necessary for error handling in some cases)
]
# Note: E402 detects module-level imports not at top of file.
# Function-level imports should be moved to top per PEP 8, but may be acceptable
# in cases like avoiding circular imports. Use with discretion.

[tool.ruff.lint.isort]
known-first-party = ["app", "api", "core", "models", "schemas", "services", "db", "utils", "workers"]
force-single-line = false
force-sort-within-sections = true
# isort will organize imports at top of file and flag misplaced imports

# ============================================================================
# Testing - Pytest
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["app", "api", "core", "models", "schemas", "services"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
fail_under = 80

# ============================================================================
# Security - Bandit
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", ".venv", "migrations"]
skips = ["B101"]  # assert statements in tests
targets = ["."]

# ============================================================================
# Commitizen (Conventional Commits)
# ============================================================================
[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version"
]

[tool.commitizen.customize]
message_template = "{{change_type}}{% if scope %}({{scope}}){% endif %}: {{message}}"
schema_pattern = "^(feat|fix|docs|style|refactor|perf|test|build|ci|chore|revert)(\\(.+\\))?!?: .+"

# ============================================================================
# Ruff Format (替代 Black)
# ============================================================================
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
